requirements.txt
boto3==1.21.28
botocore==1.24.28
click==8.1.0
distlib==0.3.4
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
filelock==3.6.0
Flask==2.1.0
future==0.18.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
Pillow==9.0.1
platformdirs==2.5.1
pyee==9.0.4
python-dateutil==2.8.2
python-ffmpeg==1.0.13
s3transfer==0.5.2
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.4
Werkzeug==2.1.0
zipp==3.7.0
Flask-Mail==0.9.1
